+++ argos/window.py
@@ -152,6 +152,7 @@ class ArgosWindow(Gtk.ApplicationWindow):
     def _setup_titlebar(self, titlebar: TitleBar) -> None:
         titlebar.central_view_switcher.set_stack(self.central_view)
         titlebar.back_button.connect("clicked", self._on_title_back_button_clicked)
+        titlebar.home_button.connect("clicked", self._on_title_home_button_clicked)
         titlebar.search_entry.connect("search-changed", self._on_search_entry_changed)
 
     def is_playing_page_visible(self) -> None:
@@ -187,6 +188,9 @@ class ArgosWindow(Gtk.ApplicationWindow):
 
     def _on_title_back_button_clicked(self, _1: Gtk.Button) -> None:
         self.props.library_window.goto_parent_state()
+
+    def _on_title_home_button_clicked(self, _1: Gtk.Button) -> None:
+        self.props.library_window.goto_home_directory()
 
     def _on_search_entry_changed(self, search_entry: Gtk.SearchEntry) -> None:
         filtering_text = search_entry.props.text